Overhauling pretty much everything to enforce compartementaliation and modularity. Making sure that mob.d only contains mob specific things, not calls to Display functions or anything else for example. Likewise with battle.d, and display.d. Thinking of restructuring Mob class and related subclasses to be super generic and not even include any named stats but only a 2D matrix of numerical values named with a 2D enum matrix. While also creating interfaces for specific functionality.
This way the engine could be easily expanded to be used for more than just turn based games and RPGs, yet still be compatible with itself. Making it easier to switch between gameplay types in game, such as inserting real time segments into an otherwise turn based game. Might make it an abstract class as part of this.

Will be creating more git branches to compartamentalize work, specificlly one for Display and one for the Combat system, and probably further subdivided. I should not be working entirely in the Unstable branch for everything.