Added two unit tests for statDisplay function and simplified frame update system in display.d. Goal for next stable release is to display 10 lines of battle log, whether empty or full, plus the names and health, stamina and mana values of heros for every frame update during a battle. We can work on fancying it up from there. No plans yet to integrate with Espeak or GTK-D.
Planning on splitting the repository in two. One repository for the as yet unnamed logic engine which I'll call the "Open Darkness Engine" for now for simplicity, and one repository for the game "Open Darkness" itself, which will be  built with that engine. Hopefully this will make things easier to work on via goal and code compatementalization.