Creating USON union as output for JSONConvert()
Because the point of JSONConvert is for it to take in a parsed JSON string and automatically format the data in each field to it's proper, static type, I needed a single return type that could output everything so I am testing out a union. The idea is that when you use JSONConvert() you as the programmer do not have to specify what sort of output it is at all, you just need to know what the data at that point should be if you don't want an error.

Example usage I'm imagining would be something like this:

string A = JSONConvert(data["A"]);
string[] B = JSONConvert(data["B"]);
int C = JSONConvert(data["C"]);
uint D = JSONConvert(data["D"]);
int[] E = JSONConvert(data["E"]);

or more generically:

auto A = JSONConvert(data["A"]);
auto B = JSONConvert(data["B"]);

etc

This module and its requisite function(s) are intended to let people define and implement their own stuff for their own project in a more user friendly and non technical way. Eventually it is intended to be integrated with a GUI tool to letnon technical people build complex things quickly and easily.