Removed body of attack() in combat.d
I want the combat system to be totally separate from the rest of the application. For example I want a person to be able to take the exact same attack function, which would define the exact stat rolls necessary to determine a characters attack, and use it in a turn based battle system just as easily as a open world real time adventuring system. It should be just a function call as far as the programmer is concerned.

I'm thinking right now that combat.d will be the place where the combat related functions are defined, and then I will implement them individualy in different places based on individual need in order to hide the implementation from the end user programmer. That way it "just works" when called.

Other ideas I've had were to use interfaces, but then I need to tie those to something, like a battle class or mob class. I don't want to do that. I want it, like everything else, to be its own separate block that just fits together with other things. Because what if you're making a point and click adventure game that doesn't require any combat but all the other mob stats can be worked into it?

Another idea has been to make a generic function that requests a definition from whatever's calling it in order to tell it what to do. But that's just the same thing I'm doing now but in reverse. Not sure if one would be better than the other. Still need to think about all of this more.