Rewriting omni union to include type checking and enforcement
Since I want omni to be a universal data structure for explicit use with JSONConvert to make sure that the programmer doesn't need to do anything other than call the function in order to assign variables properly from raw JSON data. I need it to have a single method to return its assigned, reliable value with.