Changed TBB_Mob and TBB_Hero classes to be simpler, making Hero's only defined by their ability to gain XP and level up.
Current intent is to create a PartyManagher struct or class somewhere and use that to hold hero objects, then during battle the Battle object will take teh Mob portion of that class and use it for combat. Then XP data can be fed back to teh Party Manager if you win to deal witht the hero specific Xp gain and leveling.