Replaceed Node and Tree classes with a simple Menu struct for simplicity
I may change backn in the future, or change to something else again. But for now this seems to do what I wanted it to in my mind. It made no sense for me to have bothNode and Tree classes as really a Tree is just a bunch of Nodes. I want behaviour to be handled ina  more genric way as well, which means I shouldn't need the complexity I was originally going to put into Nodes as function pointers or delegates to behaviour can be defines elsewhere.